Assistant checkpoint: Fix teams query in bowler form
Assistant generated file changes:
- client/src/components/bowler-form.tsx: Fix teams query for selected leagues

---

User prompt:

I ran it successfully but it didn't fix the problem. A couple of days ago we updated the code so we could add one bowler to multiple leagues. In doing so, did we mess up the ability to add bowlers to teams? When I go to create a new bowler and I pick a league, no teams are in the "Select a Team" dropdown, even though there are teams in the league