Assistant checkpoint
Assistant generated file changes:
- server/routes/payments.ts: Update delete endpoint to use storage interface with logging

---

User prompt:

Ok. I've restarted the server to make sure we are importing all the changes we made. I tried deleting a payment again and it didn't delete and the console logs don't show Drizzle's logs. Is it possible there is a permissions issue? We don't haver permission set up properly in the app to be able to delete transactions?